third_party/WebKit/LayoutTests/imported/wpt/html/semantics/embedded-content/the-iframe-element/cross_origin_parentage-expected.txt

This is a testharness.js-based test.
FAIL Check the frame heriarchy assert_equals: expected 200 but got 0
FAIL Check the frame heriarchy 1 assert_equals: expected 200 but got 0
FAIL Check the frame heriarchy 2 assert_equals: expected 200 but got 0
Harness: the test ran to completion.
